+++ php/7.2/entrypoint.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 set -e
 
-# With env variable WITH_XDEBUG=1 xdebug extension will be enabled
-[ ! -z "$WITH_XDEBUG" ] && docker-php-ext-enable xdebug
+# With env variable WITHOUT_XDEBUG=1 xdebug extension will be disabled
+[ ! -z "$WITHOUT_XDEBUG" ] &&
+export PHP_ZEND_EXTENSION_DIR = $( php -i | grep extensions | awk '{print $3}' ) &&
+sed -i "s/zend_extension=${PHP_ZEND_EXTENSION_DIR//\//\\\/}\/xdebug.so/;zend_extension=${PHP_ZEND_EXTENSION_DIR//\//\\\/}\/xdebug.so/" /etc/php.ini
 
 # Provide github token if you are using composer a lot in non-interactive mode
 # Otherwise one day it will get stuck with request for authorization
@@ -20,13 +22,12 @@ fi
 # Also you can set the php timezone with direct setting it in php.ini 
 # within your .gitlab-ci.yml like
 # before_script:
-# - echo "America/New_York" > /usr/local/etc/php/conf.d/timezone.ini
+# - sed -i "s/;date.timezone =/date.timezone = America\/New_York/" /etc/php.ini
 
-if [[ ! -z "$TIMEZONE" ]]
-then
+if [[ ! -z "$TIMEZONE" ]]; then
   echo "$TIMEZONE" > /etc/timezone
-  dpkg-reconfigure -f noninteractive tzdata
+  cp /usr/share/zoneinfo/$TIMEZONE /etc/localtime
+  sed -i "s/;date.timezone =/date.timezone = ${TIMEZONE//\//\\\/}/" /etc/php.ini
 fi
-echo "date.timezone=`cat /etc/timezone`" > /usr/local/etc/php/conf.d/timezone.ini
 
 exec "$@"